ata still borked, fixed a few issues.

+++ include/os/kernel/drivers/ata_pio.h
@@ -5,8 +5,8 @@
 #include <stdint.h>
 
 void ata_pio_install(void);
-int ata_pio_read(size_t lba, uint8_t *buf);
-void ata_pio_write(size_t lba, const uint8_t *buf, size_t d_len);
+int ata_pio_read(uint16_t lba, uint8_t *buf);
+void ata_pio_write(uint16_t lba, const uint8_t *buf, size_t d_len);
 void wait_400ns(void);
 
 #endif
